Refactor and import navigation
