+++ package.json
@@ -9,10 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.18",
     "expo": "~51.0.28",
     "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
-    "react-native": "0.74.5"
+    "react-native": "0.74.5",
+    "react-native-screens": "3.31.1",
+    "react-native-safe-area-context": "4.10.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
